+++ src/Faforever.Qai.Discord/Utils/Bot/CommandHandler.cs
@@ -10,7 +10,8 @@ using DSharpPlus.Entities;
 
 using Faforever.Qai.Core.Database;
 using Faforever.Qai.Core.Structures;
-using Faforever.Qai.Discord.Core.Configurations;
+using Faforever.Qai.Core.Structures.Configurations;
+using Faforever.Qai.Discord.Core.Structures.Configurations;
 using Faforever.Qai.Discord.Utils.Commands;
 
 using Microsoft.Extensions.Logging;
@@ -42,11 +43,11 @@ namespace Faforever.Qai.Discord.Utils.Bot
 				using QAIDatabaseModel model = new QAIDatabaseModel();
 				// Need to know how we are accessing the database!
 
-				GuildConfig guildConfig = await model.FindAsync<GuildConfig>(msg.Channel.GuildId);
+				DiscordGuildConfiguration guildConfig = await model.FindAsync<DiscordGuildConfiguration>(msg.Channel.GuildId);
 
 				if (guildConfig is null)
 				{
-					guildConfig = new GuildConfig
+					guildConfig = new DiscordGuildConfiguration
 					{
 						GuildId = msg.Channel.GuildId,
 						Prefix = _config.Prefix
@@ -98,7 +99,7 @@ namespace Faforever.Qai.Discord.Utils.Bot
 			}
 		}
 
-		public async Task<int> PrefixResolver(DiscordMessage msg, GuildConfig guildConfig)
+		public async Task<int> PrefixResolver(DiscordMessage msg, DiscordGuildConfiguration guildConfig)
 		{
 			if (!msg.Channel.PermissionsFor(await msg.Channel.Guild.GetMemberAsync(_client.CurrentUser.Id).ConfigureAwait(false)).HasPermission(Permissions.SendMessages)) return -1; //Checks if bot can't send messages, if so ignore.
 			else if (msg.Content.StartsWith(_client.CurrentUser.Mention)) return _client.CurrentUser.Mention.Length; // Always respond to a mention.
